Reduce state updates. Refactor code move udeCallback to separate functions.
