+++ packages/extension/src/utils/shouldUpdateState.ts
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * External dependencies;
+ */
+import { diff } from 'deep-object-diff';
+
+/**
+ * This is a utility function that decided whether or not we should update the state.
+ * @param previousState The previous state of the state object.
+ * @param update The new state of the state object
+ * @returns boolean Whether or not to update the state.
+ */
+function shouldUpdateState(previousState: any, update: any) {
+  return Object.keys(diff(previousState, update)).length > 0;
+}
+
+export default shouldUpdateState;
